refactor: move schemas to types file and remove duplicated types
